django/requirements.txt
django==5.2.4
psycopg2-binary==2.9.10
djangorestframework==3.16.0
gunicorn==23.0.0